core: 1620: add Subs class to represent unevaluated subs expressions.
This commit includes the full implementation of the Subs() class, along
with tests and pretty printing. It also includes the implementation of
utilities.iterables.uniq() function.

Subs class
==========
Subs() receives 3 arguments: an expression, a variable or list of
(unique) variables and a list of evaluation points corresponding to
those variables (must have the same length). There's no automatic
expansion - use the method .doit() for that.

The variables may be expressions, but they are subjected to the
limitations of subs().

When evaluating derivatives at a point that it's not a symbol, a Subs
object is returned. One is also able to calculate derivatives of Subs
objects - in this case the expression is always expanded (for the
unevaluated form, use Derivative()).

The objects support advanced comparison: they can be created with any
symbols as variables, with different orders of the variables and still
evaluate to equal.

Examples
--------
Simple stuff:
    >>> Subs(f(x), x, 0)
    Subs(f(_x), Tuple(_x), Tuple(0))
    >>> _.doit()
    f(0)
    >>> Subs(f(x, y), (x, y), (0, 1)).doit()
    f(0, 1)

Expressions as variables:
    >>> Subs(f(x**2), x**2, 0)
    Subs(f(_x**2), Tuple(_x**2), Tuple(0))
    >>> _.doit()
    f(0)
This won't work:
    >>> Subs(f(x), x**2, 0)
    Subs(f(x), Tuple(_x**2), Tuple(0))
    >>> _.doit()
    f(x)

Derivatives:
    >>> f(g(x)).diff(x)
    D(g(x), x)*Eval(D(f(_arg1), _arg1), Tuple(_arg1), Tuple(g(x)))
    >>> _.subs(f, sin).doit()
    cos(g(x))*D(g(x), x)
Derivatives of Subs objects:
    >>> Subs(y**2*f(x), x, 0).diff(y)
    2*y*f(0)
It also works if the diff variable is contained in the evaluation point
too:
    >>> e = Subs(y**2*f(x), x, y)
    >>> e.diff(y)
    y**2*D(f(y), y) + 2*y*f(y)
    >>> e.doit()
    y**2*f(y)
    >>> _.diff(y)
    y**2*D(f(y), y) + 2*y*f(y)

Comparisons:
    >>> Subs(f(x), x, 0) == Subs(f(y), y, 0)
    True
    >>> Subs(y*f(x), (x, y), (1, 2)) == Subs(x*f(y), (y, x), (1, 2))
    True
    >>> Subs(y*f(x), (x, y), (1, 1)) == Subs(y*f(x), (y, x), (1, 1))
    True

Operations:
    >>> Subs(f(x), x, 0) + Subs(f(y), y, 0)
    2*Subs(f(_x), _x, 0)

Pretty-printing:
    >>> Subs(f(x).diff(x), x, 0)
    /d       \|
    |--(f(x))||
    \dx      /|x=0