Delete S.Pure
See issue 2304.  Pure was a global dummy symbol that was used as the
generator of polynomials returned by minimal_polynomial() and
primitive_element().  However, it's better to just require a generator
in those functions.  This is now the case.  AlgebraicNumber.minpoly()
now also requires a generator.    It was changed to internally use
AlgebraicNumber._minpoly, which is in a dummy generator.  Internal Poly
algorithms can use that if they need some attribute of the minimal
polynomial that is independent of the generator (for example, the
degree).

This also contains a complete reversion of commit
a292eaa2bdbb9ec8e765e639da0d26cd418055bb, "Use pure symbol instead of
dummy symbols in solvers/ode.py".  Pure is not necessary there, since
RootOf now properly compares equal to other RootOf's with the same
polynomial and a different generator