Fixed ProductSet.complement
Complement of a productset is the union of all productsets where at least
one of the sets has been complemented.

Also added tests to this effect

ProductSets.complement now no longer assumes order

Was calling list on a set and assuming that a certain order came out.
Now We filter for the condition instead.

fixed square.complement test