Field.gcd() and Field.lcm() give non-trivial results
This makes sympy.polys algebraic framework coherent with implementation
of gcd() and lcm() in the core, e.g.:

In [1]: Rational(2, 3).gcd(Rational(4, 9))
Out[1]: 2/9

In [2]: Rational(2, 3).lcm(Rational(4, 9))
Out[2]: 4/3

In [3]: QQ.gcd(QQ(2, 3), QQ(4, 9))
Out[3]: 2/9

In [4]: QQ.lcm(QQ(2, 3), QQ(4, 9))
Out[4]: 4/3

In [5]: primitive(2*x/3 + S(4)/9)
Out[5]: (2/9, 3⋅x + 2)

The advantage of the new implementation of gcd() and lcm() in Field
is that primitive() now returns a polynomial with denominators cleared
and content removed, no matter if the initial domain was a ring or
a field. Some domains, like RR, don't support non-trivial gcd() and
lcm() operations, so primitive() may not return any useful results
in this case.