Fixes to .is_polynomial()
This method is *not* an is_assumption method like is_real or is_odd.
You can't assume a Symbol to be a polynomial, and the algorithm to
determine if something is a polynomial in the given variables can always
determine a True or False answer.  Therefore, the following changes were
made:

- Moved .is_polynomial() and ._eval_is_polynomial() from Basic to Expr.
It doesn't make sense for general Basic objects, only for Expr objects.
- Made Expr._eval_is_polynomial() return False instead of None.  In
other words, the base, "not implemented" case is now False instead of
None.  .is_polynomial() should now always return either True or False.
- Made (x**n).is_polynomial(x) return False, where n = Symbol('n',
nonnegative=True, integer=True).  This function was being used to check
if something was a polynomial in the sense that it could be manipulated
with polynomial algorithms like quo, so an expression like that would
have caused an error.  .is_polynomial() now checks if the expression is
an exact polynomial in the given symbols (i.e., something that could be
accepted by Poly).
- Added a docstring to is_polynomial, which explains the above and also
notes that is_polynomial is a stupid function in that it does not
attempt any simplifications.