Reimplement .is_rational_function() to be like .is_polynomial()
The implementation is basically exactly the same as .is_polynomial(),
except in Pow we don't check if the exponent is nonnegative.  This is
*way* faster than the old implementation, which called .as_numer_denom()
and then .is_polynomial() on the result.  For example:

Old implementation
------------------
In [1]: a = Add(*(exp(i*x) for i in range(1000)))

In [2]: a.is_rational_function(x)
Out[2]: False

In [3]: %timeit a.is_rational_function(x)
1 loops, best of 3: 3.78 s per loop

New implementation
------------------
In [1]: a = Add(*(exp(i*x) for i in range(1000)))

In [2]: a.is_rational_function(x)
Out[2]: False

In [3]: %timeit a.is_rational_function(x)
100 loops, best of 3: 11 ms per loop

Like .is_polynomial(), .is_rational_function() does not attempt any
non-trivial simplifications to coerce an expression into a rational
function, and a note and example of this were added to the docstring.