Create a sample tox.ini file, tox.ini.sample.
Tox is a generic virtualenv management and test command tool, we use it
to run tests in different Python versions. It requires a tox.ini file in
the base directory to function. This provides a sample file which should
then be copied to tox.ini to be used. That file will be ignored by git.