Create DecrementLevel context manager for DifferentialExtension objects
This replaces the common idiom in the Risch algorithm:

DE.decrement_level()
<do stuff>
DE.increment_level()

with the context manager / with statement idiom:

with DecrementLevel(DE):
    <do stuff>

This is not only *much* easier to read, but it's more robust, since
__exit__ is always called at the end of the with statement, even if an
exception is raised.  So, for example, you can now have a raises
statement in a test that doesn't break all other tests that use the same
DifferentialExtension object.

Previously, I didn't do this because Python 2.4 doesn't support the with
statement, but this will not be merged until after the 0.7.0 release,
which will be the last release to support Python 2.4, so I went ahead
and did this.  Note that to use the with statement in Python 2.5, you
have to import with_statement from __future__.