Robustify bool returning in Relational
It was assuming that just because expr.is_number is True, that
expr.evalf() works.  But this doesn't have to hold (evalf() might not be
implemented for example).  Although the previous commit (do not make
AtomicExpr.is_number True by default) fixes issues 2461 and 2466, this
would have fixed them too.