Use more robust hashing in polyclasses (DMP and friends)
This fixes issue 2472.  Previously, it would hash the repr of self.rep.
But this caused problems when long versions of small integers slipped
in, because repr(int(1)) != repr(long(1l)).

It now converts the nested list of lists representation to a nested
tuple of tuples.  A new dmp_to_tuple() function was added to do this,
similar to dmp_copy(), and methods to_tuple() were added to DMP and ANP.