Implement the extension flag to risch_integrate() for easier debugging
Pass a DifferentialExtension object to the extension keyword arg of
risch_integrate() to use this.  Note that it will then ignore the f and
x arguments.  Also be aware that integrating modifies the
DifferentialExtension object, so after integrating, it will need to be
recreated to be used again (actually, you should only need to increment
the level back to -1, but recreating it is probably a good idea).