Use import_module() throughout the quantum module
Also skip() tests when a module is not installed or compatibile instead
of not defining it.

The main motivation behind this change, and indeed the import_module()
function, is a bug in numpy in Python 2.5 that causes the quantum module
to not work.  Therefore, numpy is disabled in the quantum module in
Python 2.5 or lower, even if it is installed. See
https://groups.google.com/d/topic/sympy/9FOPjxC0D6s/discussion and
http://projects.scipy.org/numpy/ticket/1872 for more information.