Use import_module() to import gmpy in the polys
Also use the actual version of gmpy that is required (1.03), instead of
checking if a feature works.  This makes the code simpler (because all
checks are passed to import_module()), and when the gmpy library is too
old, the warning emitted by default by import_module will tell the user
what version he has and which minimum version he needs.

We know the oldest supported version is 1.03 because the maintainer of
gmpy said so in comment 4 of issue 1881.