Create WignerD class for rotation matrix elemnts
New class, WignerD, is returned when evaluating matrix elements of the
rotation operator, which is implemented in the Rotation class. Now
Rotation.D and Rotation.d return instances of this WignerD class with
the apropriate parameters set. This class can be evaluated using the
doit() method, which employs the evaluation method previously used in
the Rotation.D and Rotation.d functions.