Sort with key instead of cmp via helper function cmp_to_key
In Python 3, cmp is no longer supported. This means that calls to
sorted(), sort and others will no longer accept a cmp argument. For the
moment, fix this by introducing a helper function cmp_to_key in
core/compatibility. In the long run, these should all be converted to
use key directly, which is also more efficient.

A call to sort in solvers/ode was using both a cmp and a key; this was
refactored to use just the cmp and then wrapped with cmp_to_key.