Stop using tuple parameter unpacking
In Python 3, tuple parameter unpacking has been removed. Avoid this by
introducing a named variable which is then unpacked as the first
command in the function. There were six cases of this total.
Additionally, a tuple was used by a lambda in one case; this is also
fixed by naming the tuple and indexing on it.