Add a Python 3 compatible cmp() function
Python 3 no longer supports the cmp() function. Introduce a wrapper
definition in core/compatibility that defines cmp() as (a>b) - (a<b)
in Python 3 and import it where required. This fails in
core/assumptions when called with a complex number, a FIXME-py3k is
added there.

A use of cmp(a,b)>0 has been replaced with just a>b in secondquant.py

Note that this is a somewhat evil patch, especially when combined with
the cmp_to_key wrapper, as it introduces a lot of unneccessary work. It
should be considered a workaround to expose other issues in Python 3
porting. In the long run, all uses of cmp should be replaced with
Python 3 idioms. There is no performance regression in Python 2.