Coupled and uncoupled states and operators
Implements coupled and uncoupled states, including logic to rewrite and
represent these states and pass from one to the other through the
rewrite and represent functions. Also improves the handling of symbolic
states for rewriting. Coupled states are specified by taking normal
states and passing j1,j2 parameters to the represent or rewrite
functions. Uncoupled states are taken as TensorProduct's of normal
states.

Implements the logic for uncoupled operators and how operators, both
coupled and uncoupled, act on the new states. Unocupled operators, like
the uncoupled states, are defined by taking TensorProduct's of normal
operators. These operators are able to then act on uncoupled states.

For all the implemented spin space coupling, it is assumed there are
only two coupled spin spaces, functions are not generalizable to higher
spin spaces.