core/tests/test_numbers: Remove 'L' literal from test
In Python 3 int has been changed to behave as long before (meaning it's
"unlimited"). This also means code like '##L' isn't supported any more.
As this is a simplified behaviour, remove the 'L' literal from our test.
It should still work in Python 2 code, but it should be considered
unsupported.