Make all Piecewise.args Basics.
ExprCondPairs with cond True are commonly used. However this upsets lots of
code (which expects .args all to be Basic).

This is a workaround that replaces `True` by a dummy variable. This is visible
to almost no code.

TODO fix this properly when issue 2531 is resolved.

sympy/functions/elementary/piecewise.py:
  (ExprCondPair.true_sentinel) New dummy.
  (ExprCondPair.__new__) Substitute true_sentinel for true.
  (ExprCondPair.cond) Return True instead of true_sentinel.