Improve Function.diff in case of ArgumentIndexError.
ArgumentIndexError can have two meanings:

1) the argument index is out of bounds
2) the requested operation is not supported at the requested index

An example of (1) would be gamma(z).fdiff(2), whereas an example
of (2) would be polygamma(n, z).fdiff(1).

Actually fdiff is not a "public" method, so it should not be required to check
for case (1). I thus changed the Function constructor to reject constructions
with the wrong number of arguments, this then ensures that fdiff is only called
with the right number of arguments.

In case (2) it makes sense to return an unevaluated derivative.
I changed Function.diff to this effect.

Both new functionalities are tested.