Implement proper expansion algorithm for G functions.
This code is needed to expand certain G functions where slater's theorem
does not apply. Consider for example meijerg([a], [], [0, a], [], z).
This can be expanded using slater's theorem (in general), and the resulting
hypergeometric function can be recognised as an incomplete gamma function.
However, when a is an integer, slater's theorem does not apply. The result is
still an incomplete gamma function ... indeed the same as one would get by
naively substitutiong a=0 in the general formula.

There does not seem to be a general way to recognise this situation (one can
take a limit in an expansion like above and get a valid result, but (1) it is
not clear how to perturb the indices to get something we can evaluate, and (2)
computing this limit can be hard). However, it was part of the plan all along
to roll an algorithm to manipulate G functions directly (instead of relying
exclusively on slater's theorem). This is what this commit adds.

The algorithm for G functions is more complicated in the sense that there are
no good criteria (that I am aware of) to decide if one index quadruple is
reachable from another, and (hence) even if they are proving this (by coming up
with a plan) is difficult.

Thus the meijer g-function expansion algorithm differs from the one for
hypergeometric functions in three key points:

(1) The only invariants used are the signature of the index quadruple
    (i.e. m, n, p, q).
(2) Reachability is decided by hand-written functions, one for each formula.
(3) Planning is heuristic.

Some of these can be improved upon, but for now the required work does not
seem to be justified by the results.

The extensions for meijer g functions are described in detail at the top of
the file hyperexpand.py.