Remove the math_dollar Sphinx extension
This extension converted $math$ to `math` so it would render as LaTeX in
Sphinx documents.  There were problems with the extension (see issue
2537).

I changed all instances of $math$ throughout SymPy to `math`.  I also
changed some ``math`` in the polys code to `math`, though this is still
very inconsistant (I at least tried to be consistant within a file).
Note that two backticks (``) causes the text to printed verbatim in a
monospace font in Sphinx.