codegen: Refactor printing code
The statement "print >> f, code_lines," is translated to
"print(prototype, end=' ', file=f)" by 2to3. The ending comma in the
first commands surpresses the newline (which is added elsewhere in the
code) but 2to3 doesn't recognize this. Use f.write directly to avoid
this. The other uses of the "print chevron" syntax in the file aren't
affected as they don't have the trailing comma.