utilities/decorator: use iterable() instead of hasattr(.., __iter__)
In Python 3, str is iterable (has attribute __iter__), in Python 2 it's
not. We relied on this in utilities/decorator. This produced an error
when we tried to simpify a string (eg. integrate("x*y", x) failed with
SympifyError: "could not parse '*'"). Fix by importing the function
iterable from compatibility, which explicitly ignores strings, and using
that.