Remove \bm from the GA docs
MathJax does not support this, and anyway, there doesn't appear to be
much of a difference between a bold gamma and a regular one.  I didn't
know if it would be acceptable to add some kind fo arrow to indicate
that the gamma symbol is a vector, so I just am removing it for now.