Make some doctests Python 3 compatible
In Python 3 the __repr__ of many types has changed. In Python 2, the
built-in classes (list, tuple, set, dict...) represented themselves as
types while they are classes in Python 3. This is a problem because
doctests rely on comparing the output of code. Change to use the
ellipsis syntax in the three affected doctests, which will catch both
variants.