solvers.py: Use izip instead of zip in solve()
In Python 3, zip returns an iterator by default, instead of a list.
While 2to3 catches most of these and wraps them with list() if required,
it can get confused. Here it gets confused because there's a call to
zip()[0] which is already wrapped by list(). Fix this by using the izip
from functools which returns an iterable in Python 2 and using the
next() method. As izip is only used in this one (unlikely) case, import
it locally.