Remove iff()
This was needed to emulate Python's ternary statement in Python 2.4, but
as we no longer support Python 2.4, this is no longer needed.  One of
the uses of this was actually wrong.  I did not add a test for this, but
it was from the Python tokenize module, so I made sure the code matches
what is done there.

Also cleaned up some imports in some places.