Update hyperexpand to properly use polar numbers.
Even though all hypergeometric functions are unbranched at the origin, their
expression using named special functions can still contain branched functions
(just in such a combination that the branch contributions cancel, if a
constintent choice of argument is made).

Moreover hypergeometric functions *can* be branched at 1 (and oo), so branches
do have to be taken into account.

And of course meijer g-functions typically are branched at the origin.

All of this means that hyperexpand() has to properly use polar numbers, so that
1/(-1) is not changed to -1, etc.

This commit does this, updates the tests (where now a result with explicit branch
factors is returned), and adds a new test for two examples where working with polar
numbers is crucial.