utilities/runtests: Fix doctest runner for Python 3
In Python 3, the _load_testfile method has a fourth argument, encoding.
As we already use this, just pass it to the function when under Python
3. However, we then try to decode a str (because we specified
an encoding) which is wrong, so set the encoding to None in Python 3. We
cannot just use None as an encoding as Python 2 relies on the decode
step.

This error only appeared when running .txt doctests, as the usual
doctests use a different codepath.