utilities/runtests: Fix doctest runner for Python 3
In Python 3, the _load_testfile method has a fourth argument, encoding.
As we already use this, just pass it to the function (TODO: do this only
in Python 3). However, we then try to decode a str (because we specified
an encoding) which is wrong, so actually pass a "None" encoding (instead
of "utf-8") as the default.

This error only appeared when running .txt doctests, as the usual
doctests use a different codepath.