doc/../evalf.txt: Use ellipsis syntax for float() doctests
In Python 3, floats have a higher precision by default and it is not
possible to override it. This is a problem for doctests, which rely on
matching output exactly. As the purpose of these doctests is to show
limitations in Python 3 floats, we can't just convert them to the more
precise SymPy Floats. Instead, use ellipsis syntax in doctests to allow
both the higher and lower precision to pass.