doc/../evalf.txt: Skip some doctests related to float()
In Python 3, floats have a higher precision by default and it is not
possible to override it. This is a problem for doctests, which rely on
matching output exactly. As the purpose of these doctests is to show
limitations in Python floats, we can't just convert them to the more
precise SymPy Floats. Fix by skipping the four affected doctests.