Update code quality tests
The exception tests were updated to handle more cases better.  For
example, it now correctly gets stuff in doctests, but ignores talking
about an exception in a docstring (like "this function will raise
NotImplementedError, in which case the case is not implemented", which
previously was flagged as an old style raise statement).

Tests were added for the exception regular expressions.

I also fixed the import regular expression to handle ... from a
docstring and added more tests for the regular expression.

Finally, I made the tests only open each file once, and made the open()
block use a with statement.