Improve jn_zeros().
The function jn_zeros() computes the first `n` zeros of the spherical
bessel function jn of order `nu`, `nu` being real. It used to use a
generic mpmath solving routine, and return python floats.

This commit updates jn_zeros() to

o Use the mpmath function besseljzeros - this is now in the version of
  mpmath shipped with sympy.
o Return sympy Floats (unless computing using scypy).
o Accept an argument dps which controles the precision of the output for
  method sympy.

Also erase a test for method=mpmath (now default), and update the doctest.