physics/wigner: Explicitly cast range() arguments to int
In Python 2, a SymPy Float could be used as an argument of range(). This
is no longer possible in Python 3. There are two occurances of this in
wigner.py. In the first, a function which expected an integer (per the
doctest) got a Float and then called range() on it; in the second,
range() was called directly on Floats. Fix by explicitly casting to int.