doc/src/*.txt: Use init_printing() instead of setting the displayhook
Some .txt doctests were setting the sys.displayhook to pprint which
produced subtle errors in Python 3 (printing an extra "None" in some
cases). Instead, use the init_printing() function from
interactive.printing. It is called by default with use_unicode=false,
to match the expected output in the files.

These errors only appeared in tutorial.txt and matrices.txt, but use
init_printing() in four other files which were also setting the
displayhook.

Also skip a doctest in the tutorial demonstrating the old-style
division which no longer exists in Python 3 and modify two doctests in
matrices.txt not to expect a result of None (which wasn't strictly
correct).