Add rational simplification modulo a prime ideal
ratsimpmodprime computes a representation of a ratioanl expression `a/b`
such that the sum of the total degree of the numerator and denominator
is minimized. E.g.:

>>> ratsimpmodprime((x+y**5+y)/(x-y), [x*y**5-x-y], x, y, order="lex")
(x**2 + x*y + x + y)/(x**2 - x*y)

Which means that (x+y**5+y) * (x**2-x*y) is congruent to (x-y) *
(x**2+x*y+x+y) modulo the ideal generated by x*y**5-x-y.

The basic description of the algorithm is as follows: c/d = a/b mod G if
and only if a*d - b*c = 0 mod G. So try all polynomials c, d with total
degree less than the total degrees of a and b respectively.
The coefficients of c and d are indeterminates and thus the coefficients
of the reduction of a*d - b*c are linear polynomials in the
indeterminates. The coefficients, thought as a system of linear
equations, can be solved and if they have nontrivial solutions, a
simpler representation of a/b is found.
Afterwards it tries to lower the total degree of the numerator and
denominator recursively and returns the result.