Fix bugs in meijerint.
(1) Factors of (x+a)**n are most easily dealt with using expansion, not
    by writing them as a g-function. This often yields simpler answers.
    Introduce a new entry `hint` into the lookup table which can be
    used to signal conditions under which the entry should not be used even
    if it matches.

    For now _rewrite_single blindly follows this, a more intelligent scheme
    may be desirable in future (but perhaps not).

(2) Adapt meijerint_doc, the lookup table test, and _rewrite_single to the
    changed lookup table format.

(3) When dealing with g-functions of the form G(a*x**b), a should be polar.
    This is true in particular if a contains (say) real symbols which would
    be manipulated incorrectly.
    [e.g. sqrt(mu**2*exp_polar(1)) = abs(mu)*exp_polar(1) for real mu,
     this is /not/ what we want]

    Use polarify(lift=True) to achieve this, both when substituting
    into matched formulae, and when doing recursive transforms.

(4) Add a new test test_probability to test_meijerint.py that tests
    various common probability integrals.

(5) Adapt exist tests to now sometimes slightly different answers.