printing/printer: Raise AttributeError in a property
The hasattr() function is defined to call getattr() and see if an
exception is raised (in which case the attribute doesn't exist). This
has been changed in Python 3.2 to catch only AttributeErrors and not
all kinds of exceptions. See also issue #2608.

A property in printer.py checked for a key and so could raise a
KeyError, which now wouldn't be caught by hasattr(). Refactor to raise
an AttributeError if the key doesn't exist. This bug was caught by a
test in utilities/tests/test_pickling.