matrices.py: Raise AttributeError in a property
The hasattr() function is defined to call getattr() and see if an
exception is raised (in which case the attribute doesn't exist). This
has been changed in Python 3.2 to catch only AttributeErrors and not all
kinds of exceptions. See also issue #2608 and the previous commit.

The property D ("Dirac conjugation") could raise a ShapeError, which now
wouldn't be caught by hasattr(). Wrap the matrix multiplication with a
try/except statement that catches ShapeError and returns AttributeError
instead. Also change a relevant test to expect an AttributeError now.