update docstrings of Field.gcd and Field.lcm
The advantage of the new implementation of gcd() and lcm() in Field
is that primitive() now returns a polynomial with denominators cleared
and content removed, no matter if the initial domain was a ring or
a field. Some domains, like RR, don't support non-trivial gcd() and
lcm() operations, so primitive() may not return any useful results
in this case.