Add FGLM
FGLM converts a Groebner basis of a zero-dimensional ideal from one term
order to another. The current implementation loses some speed when
constructing the matrices which represent multiplication by a variable.
The actual conversion takes O(#variables * #solutions^3), where the
number of solutions counts multiplicity.
It uses lists of lists instead of matrices because mpqs become
sympy.core.numbers.Rational in a Matrix but mpqs are crucial for speed
as the entries in P tend to become very complicated.

Examples:
>>> F = [x**2 - x - 3*y + 1, y**2 - 2*x + y - 1]
>>> G = groebner(F, x, y, order='lex')
>>> fglm(G, 'lex', 'grlex', x, y)
[x**2 - x - 3*y + 1, -2*x + y**2 + y - 1]
>>> groebner(G, x, y, order='grlex')
[x**2 - x - 3*y + 1, -2*x + y**2 + y - 1]

To give an impression why this could be useful, consider cyclic5:
>>> a, b, c, d, e = symbols('a b c d e')
>>> F =  [a + b + c + d + e,
    a*b + b*c + c*d + d*e + e*a,
    a*b*c + b*c*d + c*d*e + d*e*a + e*a*b,
    a*b*c*d + b*c*d*e + c*d*e*a + d*e*a*b + e*a*b*c,
    a*b*c*d*e - 1]
>>> %timeit groebner(F, a, b, c, d, e, order='lex')
1 loops, best of 3: 42.2 s per loop
>>> %timeit G = groebner(F, a, b, c, d, e, order='grlex');
    fglm(G, 'grlex', 'lex', a, b, c, d, e)
1 loops, best of 3: 2.99 s per loop