Add "hyperrepresentation" functions.
For the slater expansion, we need analytic continuations of hypergeometric
functions to the riemann surface of the logarithm with very specific
properties.

Add a new class HyperRep to represent such analytic continuations, and a
number of concrete subclasses. Test numerically that the properties are
fulfilled.

Also extend the extract_branch_factor() method of Expr to optionally extract
half-periods.