Make the doctest runner test example files
A few of the example files have doctests in them, and these weren't
being tested.

This uses the same logic as python -m doctest (see the _test() function
in the standard library doctest.py file).  The example files to not have
corresponding __init__.py files, so we have to manually add the path to
sys.path before importing.  As a result of this, to import functions
from these files in the doctests, just import it from the filename.  So,
for example, if you have a function foo() in examples/advanced/bar.py,
then you would do

def foo():
    """
    Do stuff.

    Example:

    >>> from bar import foo
    >>> foo()
    output
    """
    <...>