Make str(sqrt(x)) == "sqrt(x)" (instead of "x**(1/2)")
This fixes issue 2359.  The actual change is in sympy/printing/str.py.
Everything else is just fixing things that broke with the change (mostly
doctests).  There was one non-trivial fix, and that was a custom
IntervalPrinter in the polys that had to be rewritten.  Everything else
just worked. Also, I had to rewrite a bit from the gotchas that would
have no longer made any sense.

I also grepped the entire repo for "\*\*(1/2)" and "\*\*(S(1)/2)" and
replaced all relevant instances with sqrt().  This way, the whole repo
is consistant with this.

This change makes things look *much* better, and 90% of the time, you
can just copy and paste a printed expression, and it will just work
(before, if you had any square roots, it would not without sympifying
the string because of the 1/2 in the power).