Always rewrite product(a**b) as a**summation(b)
Previously, this was only done if the summation evaluated to something
(other than a Sum()).  See issue 1718.  For example, we now get

In [1]: product(2**f(k),(k,1,n))
Out[1]:
   n
  __
  \ `
   )   f(k)
  /_,
 k = 1
2

I'm not sure if this is a good idea, so I'd be perfectly willing to
reject this patch if others don't like it.  But it seems OK to me,
especially considering how weak our product machinery is compared to our
summation machinery.  For example, Product.evalf() doesn't seem to work,
Product doesn't pretty print (see issue 2636),and Product is seriously
broken (see issue 2639).