quantum/__init__: Change imports syntax for Python 3 compatibility
The __init__.py file for quantum uses some importing magic so that both
"from quantum import *" and "from quantum import State" to work. This
fails under Python 3 because some underlying rules seem to have changed.
The likely cause is that the module names shadow function names (or
vice-versa). Fix by directly import __all__ from the affected modules. A
better fix might be to change the module names, but this is a
backward-incompatible change and this workaround is equivalent
functionally.