Fixed Trotter-Johnson ranking; cardinality; optimizations.
Fixed rank_trotterjohnson, added unrank_trotterjohnson,
added the cardinality property.

__pow__ is generalized to work also for 0 or negative power; used
binary splitting for large powers.

Used ifac instead of factorial, since it is likely that if
factorial(n) is called once it is called many times, in which case
ifac uses a cached value.

Optimized parity in the case in which the cyclic form is available.

Made some trivial optimizations; the biggest speedup is in __add__.
Instead of computing the inversion_vector each time (which is not
a property but a calculation) these values are stored and then
accessed element by element to produce the result.