external/importtools: Don't import NumPy if running PyPy
PyPy, as of version 1.6, includes some basic NumPy support. This work is
not complete and some uses of NumPy crash PyPy. This is in particular a
problem when running tests as it breaks the test suite. As NumPy is an
optional dependency, simply do not import it at all when under PyPy.
Because all external modules should be imported using the import_module
function from external/, add a check there.

This commit should be reverted once PyPy has full NumPy support.