heurisch: exhaustive search for correct substitution order
Finding a proper substitution order is critical at later stages of
heuristic Risch algorithm. Previously architecture dependent order
was used, which caused heurisch() to find an antiderivative on one
platform and to fail on another (e.g. pypy).

The new approach is to iterate over all permutations of substitution
orders and use a cheap criterion to select a valid one. If none of
the permutations were valid, the algorithm either fails or retries
with rewrite=True. The criterion isn't completely accurate, so it
may happen that the algorithm won't be able to find an antiderivative
and there exists a better substitution order. In this case, the
algorithm will resume iteration over permutations until a new,
valid substitution order is found. This process may continue until
an antiderivative is found or the space of permutations is exhausted.
This scenario is the worst case, making heurisch() really slow,
because for every valid but unsuccessful substitution order the
whole algorithm has to be run, including the linear solver part,
which is the most time consuming. The criterion itself is very cheap
and consists of a few .is_polynomial() and .is_rational_function()
method calls.