heurisch: restart the algorithm at most 3 times
This doesn't allow heurisch() to hang if there are very many valid
but unsuccessful substitution orders. A good example is an integral
from examples/advanced/gibbs_phenomenon.py:

In [1]: integrate(Abs(x**2 - S(1)/3)**2, (x, -1, 1))
Out[1]:
1
⌠
⎮          2
⎮  │ 2   1│
⎮  │x  - ─│  dx
⎮  │     3│
⌡
-1

In this case, almost every substitution order is valid and causes
the algorithm to proceed with expensive computations which don't
lead to finding of antiderivative. Previously the number of retries
was about 5! which takes oo time to compute. Now only three retries
are allowed and heurisch() gives up after a few seconds.

Note that retries are counted separately for rewrite=True/False, so
if the integrand contains trigonometric or hyperbolic functions, the
total maximum number of retries will be 6.