test_pickling: Update to test protocol 3; cleanup FIXME notes
As we now support Python 3, we should test the third version of the
pickling protocol too. As Python 2.x doesn't support it, we test it only
if we are in Python 3.

Also remove some FIXME-py3k notes about issues which were fixed earlier
(commits f732c4eb and 1dec4c5a, respectively).