Update README with information about Python 3 usage
Add a section to the README detailing how to use SymPy under Python 3.

Also change the shebang line in bin/use2to3 to use python3 by default.
While it is possible to run it under Python 2.7 too, there's no point
making Python 3 code if there is no Python 3 interpreter to run it with.