Fixed test_exp_pade, added precision in exp_pade.
test_exp_pade did not test at all exp_pade; put a condition on the
matrix for the similarity transformation to avoid using high precision.

Added precision in exp_pade depending on the order of the matrix,
because tests show that without doing this the precision of the result
decreases with the order of the matrix.