__init__: Raise ImportError in unsupported Python versions
Currently, running SymPy with an unsupported Python version results in a
SyntaxError somewhere. As this is not intuitive to users, raise instead
an ImportError. This required adding a check if 2to3 was run on the
codebase or not, which we do by checking long.__name__. We can then give
the appropriate error message. As the attribute is no longer needed
after the start, delete it. This fixes issue 2619.

Also add the bin/use2to3 script itself to the list of files not to be
copied over. This should prevent the user from accidently running 2to3
twice on the same code (which results in nonsense code).