Remove some mentions of Python 2.4 in comments and documentation
As we no longer support Python 2.4, remove mentions of it in the
comments. Some of these changed files are crufty and no longer used, but
changed nevertheless (test_pickling and particularly pytest.py). There
are a few remaining mentions: in a couple of mpmath files, tutorial.txt
and a comment in test_pretty.

The following command was used for searching: git grep "[Pp]ython\s*2.4"