bin/use2to3: Fix shebang lines in converted files
The last remaining issue for seamless Python 3 support is changing the
shebang lines after running 2to3 on the files. Per PEP 394, Python 3
code should use python3 in their shebang line. We do this by comparing
the exact string "#!/usr/bin/env python" and changing it to point to
python3, in all files 2to3 was ran on. This is not perfect, but use2to3
execution time is still dominated by 2to3 itself, so it is not a
problem. This also required changing a few files to use that exact
string (some had extra spaces after the shebang).

This also uncovered another issue: files with no extension will not get
passed to 2to3. We can work around this for the shebang fixer, and as it
doesn't cause other problems (currently), I have just added a FIXME
about this.