Raise TypeError when wrong number of args was passed to a function
Previously ValueError was raised which was semantically wrong and also
made it hard to cooperate with other libraries. For example, in NumPy
vectorize() assumes that TypeError (with specific error message) is
raised when wrong number of arguments is passed to a function. Now it
gives:

In [1]: import numpy as np

In [2]: np.vectorize(sin)
Out[2]: <numpy.lib.function_base.vectorize object at 0x2372750>