setup.py: Add clasifiers, long_description
setup.py can define metadata about the package. Some of it is
non-mandatory, but is useful for PyPi and elsewhere. As such, add a
long_description field (copied from current PyPi package, which was in
turn copied from the website) and a list of classifiers (based on the
current list on PyPi).

This fixes issue 2615.