Issue 101: Automatical tests on the wiki documentation.
This procedure test wiki pages against master or release  repositories.
Assuming that wiki.sympy and release version are placed near master repository.
E.g.
    ./sympy/
    ./wiki.sympy/
    ./sympy-0.7.1/

If the directories are not the same, then they can be tuned by
those options:

    -a AGAINST, --against=AGAINST
                    What do the tests must be relative [master | release |
                    master,release] [default: master]
     -w WIKI_DIR, --wiki-dir=WIKI_DIR
                    The path to the wiki pages directory
    -m MASTER_DIR, --master-dir=MASTER_DIR
                    The path to the sympy master repository
    -r RELEASE_DIR, --release-dir=RELEASE_DIR
                    The path to the sympy release directory

The files are to be passing through the test contain the special directives:
For reStructedText comments is shaped as:

     .. wikitest release

For markdown:

    <!-- wikitest release -->

"wikitest release" directive means that tests for current wiki-page have to be
    passed against release version only.
"wikitest master" - against master.
"wikitest master,release" - against master and release version